Provide factory method for vanilla kredis types
